Add Players, use board coords for squares, undo now undoes kinging
Realized that multiple jumps in one turn are supposed to be done with the same piece. The first jump can be with any piece able to jump, but subsequent jumps must be with that same piece. Also, if a piece is promoted, the turn ends, even if more jumps are available. These need to be added.